Refactoring of code for better structure as well as creating a copy of the client profile page for professional side
