+++ app/src/main/kotlin/diettracker/db/tables/SavedMeals.kt
@@ -1,5 +1,9 @@
 package diettracker.db.tables
 
+/**
+ * This table stores meals saved by clients and stores them in the database
+ **/
+
 import diettracker.db.MAX_LEN
 import org.jetbrains.exposed.v1.core.Table
